requirements.txt
requests
python-dotenv
streamlit
groq